Move the transaction decorator to storage
The goal is to find a better place for the transaction decorator and
adjust the implementation to allow disabling the decorator. This will
allow testing transactional methods in isolation without having to mock
out all the code that tests retries and handles transactions.

Here is an example of disabling the transaction:

   @mock.patch('designate.storage.transaction.ENABLE_TRANSACTION', False)
   def test_method_in_isolation(self):
       assert self.central_api.create_zone()

You can also explicitly enable or disable the use of a transaction or
the retries.

Change-Id: If3c33940679b41bb09e0d8b92f46f757cba78e8e